Move getEstimatedGasLimit to shared libs
