--- apps/bond/common-util/functions/requests.js
@@ -1,16 +1,0 @@
-const ESTIMATED_GAS_LIMIT = 500_000;
-
-/**
- * function to estimate gas limit
- */
-export const getEstimatedGasLimit = async (fn, account) => {
-  try {
-    const estimatedGas = await fn.estimateGas({ from: account });
-    return Math.ceil(estimatedGas * 1.2);
-  } catch (error) {
-    window.console.warn(
-      `Error occurred on estimating gas, defaulting to ${ESTIMATED_GAS_LIMIT}`,
-    );
-    return ESTIMATED_GAS_LIMIT;
-  }
-};
