+++ apps/govern/common-util/functions/requests.ts
@@ -1,5 +1,4 @@
 import { readContracts } from '@wagmi/core';
-import { Contract } from 'ethers';
 import { AbiFunction } from 'viem';
 
 import { STAKING_FACTORY } from 'libs/util-contracts/src/lib/abiAndAddresses';
@@ -10,29 +9,6 @@ import { Address } from 'types/index';
 import { getAddressFromBytes32 } from './addresses';
 import { getStartOfNextWeekTimestamp } from './time';
 import { getVeOlasContract, getVoteWeightingContract } from './web3';
-
-const ESTIMATED_GAS_LIMIT = 500_000;
-
-/**
- * function to estimate gas limit
- */
-export const getEstimatedGasLimit = async (
-  fn: Contract['methods'],
-  account: Address | undefined,
-) => {
-  if (!account) {
-    throw new Error('Invalid account passed to estimate gas limit');
-  }
-
-  try {
-    const estimatedGas = await fn.estimateGas({ from: account });
-    return Math.ceil(Number(estimatedGas) * 1.2);
-  } catch (error) {
-    window.console.warn(`Error occurred on estimating gas, defaulting to ${ESTIMATED_GAS_LIMIT}`);
-  }
-
-  return ESTIMATED_GAS_LIMIT;
-};
 
 type VoteForNomineeWeightsParams = {
   account: Address | undefined;
