+++ libs/util-functions/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/util-functions';
 export * from './lib/sendTransaction';
 export * from './lib/sendTransaction/helpers';
 export * from './lib/notifications';
+export * from './lib/requests';
